qfileinfo - revert suffix() behaviour to Qt 4.7
This workarounds ugly bug when icons in QtMoko have invalid size.
This is because suffix() now returns empty string for resouce files
like :/icon/play - it should return .pic

Because of invalid suffix wrong iconengine is picked up - it should
be qtopiapicengine and also suffix is used in
QThemeImageItem::resizeImages to detect pic format.

Bug report is here:

https://bugreports.qt-project.org/browse/QTBUG-27344